修正设备序列号生成位数不足的bug, 定时器冲突对比时,不区分类型.

+++ Thermostat/Model/LinKonDevice.m
@@ -81,8 +81,8 @@
  @return 序列号
  */
 - (long long)randomSN {
-    long long tempSN = (rand() % 9000 + 1000) * 10000 * 10000 * 10000;
-    tempSN += (rand() % 10000) * 10000 * 10000;
+    long long tempSN = ((long long)(rand() % 9000 + 1000)) * 10000 * 10000 * 10000;
+    tempSN += ((long long)(rand() % 10000)) * 10000 * 10000;
     tempSN += (rand() % 10000) * 10000;
     tempSN += (rand() % 10000);
     return tempSN;
